requirements.txt
quart==0.19.6
botbuilder-core==4.15.0
botbuilder-schema==4.15.0
python-dotenv==1.0.1
openai==1.23.6
Office365-REST-Python-Client==2.5.11
pandas==2.2.2
openpyxl==3.1.5
aiohttp==3.10.0
asyncio==3.4.3
pydantic==2.8.2
uuid==1.30
git+https://github.com/ingrammicro-xvantage/xi-sdk-resellers-python.git@6d1e7ea944babbcf085fffe9f9046458275cb0c9